pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "macro_regime"
version = "0.1.0"
description = "Regime-based macroeconomic analysis project"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "scipy",
    "pandas>=2.0",
    "numpy>=1.24",
    "fredapi>=0.5",
    "yfinance>=0.2",
    "requests>=2.28",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]